web_server/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
requests==2.24.0
riotwatcher==3.1.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.5.1
